fix: delete diskmanager_supported
When the change to the logger methods to require a context was made,
this file was missed. For the majority of the time that is okay. However
it fails compiling on an operating system other than Linux.

However the better tack is to delete it. We only support juju
agents on Linux.