refactor: make new unit,machine exist funcs
Previously in the model state we needed to check if a machine or unit
existed and to do this we piggy backed off of the is dead call and
discarded the dead aspect of this work.

This didn't read well and was bit lazy. This commit adds two new checks
to see if these entities exists and migrates the calls over to these
helpers.

This was done as PR feedback.