chore: fetch machine host key
When dialling the machine, we verify its host key by using the keys known to the Juju controller.

To avoid ambiguity between the number of structs/methods that include the term host-key, I've renamed the sshserver controller facade methods and structs to explicitly mention the word "virtual".

The client facade for `sshclient` has the same issues but I have not changed the naming of the facade there to avoid backwards compatability issues.