fix: always create k8s unit virtual host keys
Virtual host keys for k8s units were only created when scaling the application because the `newUnitVirtualHostKeysOps` method was in the wrong place.
app.AddUnitOps() is only called when scaling a unit.
On application creation, the more general `addUnitOpsWithCons` is called so we move the host key ops logic there.