refactor: remove placement from import
We want to be very explicit about when importing that you're
importing a machine for IAAS. We don't want to mention placement,
as it gives the wrong impression.

Future work will require bifurcating the import application for
IAAS and CAAS, so we can drop the if statements. That's out of
scope for this PR.