refactor: model http context injecting
This change adds a new function to http context for helping to set the
model uuid on a http request context. Previously this was not easily
done and was spread over two functions. By putting it into one function
we can control how this is done in one place and also allows for easier
mock testing in the future.

While here the oppurtunity was taken to use the concrete type for model
uuid.