refactor: introduce new object store interface
Until now we have had two object store getter interfaces. A generic one
that lets you get an object store for an arbitrary namespace that the
user supplies and a model scoped object store. Essentially the model
object store is just scoped to a singular namespace.

Where this becomes confusing is we have the concept of a contnroller
wide object store that is not tied to any particular model. However when
we pass the getter around for this we use the ModelObjectStoreGetter
interface type.

To the modest reader this conveys that the object store might somehow be
linked to the controllers model when it is not. This commit introduces a
new NamespacedObjectStoreGetter that is fixed on to a singular namespace
for the requestor.

The model getter is now just an alias for the namespaced version.