feat(unit): add unit watcher for uniter
The uniter worker uses a watcher provided by the WatchUnit facade
endpoint, which emits changes when _any_ property of a unit's doc
changes.

Unfortunately, this is a little problematic to port over to DQLite,
since these properties are scattered over a few tables, and it is not
clear which are important to watch.

However, at the moment, we cannot do much about this, since we need to
maintain facade compatibility. Leave a TODO to resolve this when we can
bump the facade again. Ideally, we would provide many specific watcher
to the uniter, rather than once generic one.

Inspecting the uniter, we can see that the triggers it needs are:
- changes to the unit's resolve mode
- changes to the unit's life
- changes to the unit's charm
- changes to the unit's subordinates

We can cover all these by watching three tables. (unit, unit_principal,
unit_resolved)

Implement this watcher on the application service