refactor: add agent binary store methods
This commit refactors the add methods for agent binaries  in the store.
We never had a specific error to represent the fact that a sha mismatch
has occured. Instead we were relying on the object store error that
wasn't documented.

This change explicitly calls out this user problem now as a first class
error. This was done so that the not valid error used for a bad agent
binary version can be told apart.