refactor: logging for agent binary store add
When uploading new agent binaries to an agent binary store via the tools
upload handler we wanted a way to log explicitly in the controller that
this operation is happening and who it is being performed by.

This is so eventually we can get to point of accountability of how a
given controller got into the state that it is in.

The first attempt had this as a log message inside of the upload
handler. The main problem identified with this approach is that this requires
the handler to understand the concept of authn and what that means. Even
though it isn't the responsability of the handler to deal with
authentication and authorisation.

This responsability falls to the register of the handler and what
control mechanims they wish to place around this.

The other issue with the log messages is it wasn't able to correctly
identify the agent binary store that was being uploaded to. We couldn't say
explicitly this model or that controller.

To deal with the above a wrapper has been made around the agent binary
store that carry's the context of the request including the name for the
store. With this we can produce a log message with enough context. This allows the handler to
focus on business logic and not that if it's dependencies that it has no
direct control over.

As part of this commit I have also introduced the ability to get an
entity from a context. What this means is that in the apiserver if we
can reasonably identify the entity that started the context for the
request we will provide a string to represent the entity.

The term entity is purposely left vague so that the caller doesn't try
and interpret the value but uses it verbatim as data for someone to
read.