refactor: migration master tool collection
This commit changes the migration master facade and client to now
consider tools with regards to their sha256sum. This is done as part of
model migration where the migration master performs a final step of
uploading agent binaries from the source controller to the target
controller.

Until now we have always downloaded the binaries from the source
controller using the agent version however there exists no garuntees
that this is what the agents in the model are currently running or need.

It is much more safe and efficient to fetch the binaries based on the
sha256 sum. This way we are uploading exactly what is required and not
potentially the wrong thing.

This also helps bolster our correctness in model migration reducing the
risk of security issues.

I haven't bothered bumping the facade version for this change as it is a
controller facade and we are looking to eventually migrate off of this.