fix: flaky tests in sshserver worker
Closing the listener used in the SSH server outside of the server's Close() method doesn't allow the server to know it's in a "dying" state so the error returned by Serve() is sometimes an unexpected one, causing flaky tests.

This change isn't strictly the most correct fix but an ideal fix would be a change to the upstream Gliderlab's SSH library.
For now, since the error returned by Serve() is not useful to us anyway, we can just log it and always return the tomb's ErrDying error instead.