feat(resolve): add new purpose-built resolve facades
Previously, to detect weather a unit's error state has been resolved,
the uniter would detect if _any_ change has been made to the unit,
and refresh a bunch of things in it's cache.

This is very messy, and doesn't transfer over to DQLite very well.

Instead, expose a new watcher to the uniter which watches for specific
changes in the unit resolve mode, and use this to update the cache
instead.

This is the start of a process to de-couple watchers in the uniter from
eachother, and to deprecate the generic WatchUnit facades.