fix: allow implicit relation supplied by juju
encodeMetadata is called for newly downloaded charm metadata and charm
metadata read from the database. Due to RI, the formerly implicit
relation endpoint `juju-info`, as requires, is now added to every
charm which does not have an endpoint with the name `juju-info`.

This change ensures that verifyNoReservedNameMisuses allows for a
`juju-info` endpoint as a required role. Other wise using a charm for
more than one application will fail.