refactor: tools agent binary shim
Refactors the agent binary shim for a model in tools to error when the
request doesn't contain a model uuid. This shim is only supposed to be
used for model requests so it makes sense that an error should be
produced if this information is missing from the request.

This was in response to PR feedback.