feat: bulk get agent binaries for machines and units
When exporting machine and unit tools via model export we require the
ability to get what "tools" or agent binaries are running for these
entities.

This information is a hybrid of what the entity has reported to the
controller that it is running and also what agent binaries have been
registered in the model's agent binary store. We have never had a direct
correlation between this information in Juju as we don't store any
information around what these entities were provisioned with.

For these calls to work we must have had the entity report its agent
version and tools stored in the store for them.

If neither of these cases are true then these calls will fail until we
can work a better design into the system.

The new state functions added in this commit also supply this
information in bulk to the caller for easier exporting. This allows us
to not tie up the database querying for each enitity in a model when
there could be 100s.