chore: improve test consistency for ssh server max connection
The test TestSSHServerMaxConnections that tests the server's maximum concurrent connection limit displayed flakey behaviour. This is likely because we are relying on server side logic from the client without any timing guarantees.
This change validates the server's atomic count of connections at each important step to make the client and server timing consistent.