chore: refactor proxy handlers
Move the k8s and machine proxy handlers from a single file into a separate package, with a dedicated file for each type of handler, similar to the Gliderlab's library.
Also introduces a factory pattern for creating the appropriate type of handler based on the destination - k8s or machines.