feat(relation): create subordinate on enter scope
This change adds code to create a subordiante unit on enter scope when
needed. A new function is added to the state layer to check if a new
subordinate is needed, and return the parameters for creating it if it
is. This is called as part of the EnterScope service method.

The EnterScope service method now also takes a closure which is expected
to create a subordinate unit. In a follow up PR, a method from the
application domain will be passed in here to create the subordinate.
This allows the code from the application domain for unit creation to be
re-used while remaining the concern of the service method that makes use
of it.

Note that in 3.6, the creation of subordiante units was done in the same
transcation as EnterScope. This is not the case here. This has been
noted in the risk register.