fix: allow container placement
The placement of a scope container can use the colon separator to
place a unit onto a machine whilst targetting a machine: lxd:0.
Thus inserting a unit on to any lxd container on machine 0.

The code error'd out originally, preventing this, but there are no
unit tests in both core or state that permitted this. It was just
transient knowledge.