feat: remove state from bootstrap
The state package was used in the bootstrap worker to dual write
applications and units. We need to relieve ourselves from that
burden and allow the controller to come up without any state
involvement; minus the machine access.

The code is pretty simple, remove the dual writing. I think we'll
be luckily if this comes up first time and no other work will be
required. If that's the case we'll need to fix those other call
sites as well.