feat: add unit address watcher
This patch adds a new watcher to the application domain that watches for
changes in the unit's or cloud service's address' or endpoint bindings
changes.
This watcher is intended to be used in the WatchUnitAddressesHash uniter
facade.

The change in behavior with respect to the legacy state is that we now
only have one watcher that watches for addresses and endpoint bindings
change for units and cloud services at the same time, without having to
bifurcate for k8s and machines.

As a side quest, the watchertest Check method was updated so that it
reports the error returned by the watcher. This is useful when we want
to make sure that an initial query does not fail (before this change we
would simply have a failed assertion because the changeset is empty).