refactor!: remove status filtering
Status filtering should be tackled in the status epic. There are
lots of problems with the current filtering architecture and
design. Notably;

 - It's post database request. For every request of the status, we
   retrieve _every_ single item and then filter it out afterwards.
   Filtering should reduce the number of items we should be able
   to retrieve from the database.
 - The filtering logic is hand tied by the state structs. Converting
   them to be generic interfaces or local types is going to slow
   the migration to dqlite right down.

Removing the patterns breaks the command show-machine (there might
be others). The **fact** that these commands are recalling the
**WHOLE** database for **ONE** machine is insane. This is terribly
inefficent.

Remove it all, we can implement it later.

BREAKING CHANGE: Removes status filtering.