feat: add public and private address retrieval for units
This patch adds the GetPublicAddress and GetPrivateAddress service
methods to the application domain, which retrieve the public and private
addresses for a unit name.
This methods are only used for k8s provider for the moment, because in
the future we'll also retrieve the machine addresses inside these
methods. But we need to finish migrating machines to dqlite first.