feat: state methods for relation migration
AddRelationWithID does almost the same thing as AddRelation, but defines
the relation ID ahead and does not require the endpoint return values. A
relation uuid is returned for next steps within the relation service.

GetApplicationIDByName is a helper method for the relation service only.

DeleteImportedRelations clears data added by the migration of relations.
Part of a rollback scenario.