feat: import of relations during model migration
Follows the precedent of other domains which call normal service and
state methods to introduce the migrated data.

There are 3 caveats: relation id sequence is is migrated in the sequence
domain; peer relations are migrated in the application domain as they
are part of CreateAppliation; relation status is migrated as part of the
status domain.