feat(application): import subordinate information
This change imports the information about unit principals. If a unit has
a principal unit, that information is recorded in the unit_principal
table.

Note that the principal is noted rather than the subordinates of a unit
because the subordinates are imported after the prinicipals. If we try
and record the subordaintes of a unit, they might not exist yet.