refactor: controller only agent binary store
The agent binary store is only saved in the controller database.
There isn't a concept of model only binary stores. This would break
a lot of assumptions of how we store our binaries.

This replicates how 3.x works and I don't think we want to change
that atm. It does mean that we could potentially have one binary
across all models on disk - which is a huge win!