fix: setup and use closure for creating peer relations in application
The CreatePeerRelationsFunc allows for a closure to be passed to
CreateApplication in the providerService of the application domain for
creating peer relations. This is the second step to migrating peer
relations with the correct relation id.

There is risk involved here as it's possible, though unlikely, that an
application can be created while its peer relations fail to be created.
There would be no way to force this after the fact other than upgrading
the charm if a new one was available. This differs from the strategy in
juju 3.x.

CreateApplication in the migrationService of the application Domain has
no need to create peer relations as their migration will be handled
inside of the relation domain.