feat: retrieve all unit addresses in one single sql query
This patch updates the public/private address state layer with only one
single query for both cases. The scope matching logic is kept in the
service layer, as it was previously done in the legacy state using core
methods.
Also, the k8s pod addresses are inserted with the MachineLocal scope, as
it's done in the legacy state as well.