refactor!: remove available to from schema
This removes the AvailableTo from the schema gen. It was originally
requested by the web team for the dashboard to discern if they
could call an API method. The original logic was written by me, but
it was based on flawed assumptions, because people didn't correctly
write or group methods in a facade that didn't violate preconditions.
For example; API facade methods were put on facades that could
violate the assumptions of the facade. It was also possible for a
facade method to dip into different model state and even the
controller state, which would violate if a API method is for a
particular model or controller.

If this is still required, it would require to be manually created
and looked up against each method.

BREAKING CHANGE: Removal of the AvailableTo from the schema. Work
around is to inspect each method and assess if it's for a controller
or a model.