fix(relation): resolve datarace in watcher relation handling
This patch fixes a datarace caused by concurrent access to the `relationsIgnored` set in
the relation watchers.

Both filter and mapper of the watcher were using a set to discard some ignored relation.
- Filters were using this set to actually ignore some relations
- Mapper was using this set to update ignored relations.

The issue is that filters and mapper are run in two separate go-routine, which cause the data race.

In this patch, the incriminated set has been moved from the watcher struct and inlined in the
GetMapper function.
It should prevent misuse but implies that a "mapper" is stateful and shouldn't be call in
eventstream package as `GetMapper()(ctx, txn, changes)` (since it would break the state).

I checked, it is the case, and doing it as above would be a smell.

- `domain/relation/service/watcher.go`:
  - Removed shared `relationsIgnored` field from `principalLifeSuspendedStatusWatcher` and `subordinateLifeSuspendedStatusWatcher`.
  - Moved `relationsIgnored` to a local variable in `GetMapper`, isolating its usage within the
    Mapper lifecycle.
  - Updated `GetFirstFilterOption` and `GetFilterOptions` to replace `PredicateFilter`
    with stateless `NamespaceFilter`.
  - Improved concurrency safety and eliminated potential inconsistencies in relation filtering
    logic.
- `domain/relation/service/watcher_test.go`:
   - Test updated due to the inlining of `relationsIgnored` field.