feat: add-ssh-jump-cli-option
Add juju ssh --jump option. With this option you can use the ssh jump
server using the juju CLI. What it does is: resolve the target into a
virtual hostname, fetch jump host key, check between the controller
addresses the one that is reachable, and setup the proxy command using the
jump serverssh. The structure is the exact same of the other providers:
ssh-container and ssh-machine + some methods that aren't needed anymore
because we have less logic in the CLI.