fix: relationunitwatcher data race
The test for the relation unit watcher had a data race. The
notStartedSafeGuard channel was set to nil whilst the pointer could have
been in the process of being read by another go routine.

The notStartedSafeGuard was intended to be used as a safeguard to avoid
leaking the go-routine that consumed the watcher events in the case of
the watcher never starting. This was done by intially leaving the
channle open, setting it to nil when the watcher started, and then
closing it with a defered finisher (if it was still not nil) at the end
of the test. When the channel was nil, it would never be selected, but
as soon as it was closed, all recivers would get a nil event. This would
allow the consumer go-rountine to exit (with an error).

The race here was that the channel was set to nil when the event
consumer started, however, the producer could still be reading the
pointer in the channel variable.

It appears that the nil channel was primarily being used as a signal to
the finisher, and that an open channel with no events being sent down it
would also suffice in this case.

To fix this, I have used a boolean signal to indicate weather the
watcher was started. This is not touched from within the go-routines,
and allows the finisher to decide weather to free the producer
without worrying about weather the channel is nil or not.