refactor: change names of applicaiton settings funcs
There were two functions to set the application settings, one for the
remote application and one for the local application. For the remote
application leadership is required, otherwise, these two functions are
identical. They both take the ID of the application to update, rather
than the applicaiton calling them. Change the names to reflect this.