feat: adds the ability to check if a model exists
Adds the ability to the model service otherwise known as the model
manager to assert if a model exists or not. This was added in response
to some changes made to the apiserver where we were calling other
service methods to get model info and relying on the error return.

Ideally we don't want to do this as it produces data we don't use and
errors we have to translate.