refactor(action): remove final state application from the action facade
The EnqueueOperation facade endpoint used FindEntity from Mongo state to
retrieve a unit or a machine from Mongo state. As we finalise cutting
over to the application domain, this will break.

Refactor the EnqueueOperation facade to drop the call to FindEntity and
replace it with a NotImplemented error.

This is the least invasive change.

However, this means that actions will be broken until we complete the
actions domain.