feat: add origin to space address
This patch adds address origin to the SpaceAddress.
This is necessary to change sort order of the space addresses, which is
needed to ensure that unit (public/private) addresses get sorted
correctly and therefore have idempotent results.
As a side change, the "host" origin was changed to "machine" in the
network DDL, to match the constant we have in core/network/origin.