feat: wire up unit workload version
The workload version is used by charms to indicate the version of
the workload. It was previously stored in the status collection as
a way of reusing infrastructure and to prevent units and/or
applications waking up additional watchers.

Instead, this change, just adds the information independant tables,
that won't wake up watchers that aren't looking for the changes on
those tables (they don't have any triggers).

The application workload version contains the denormalized last
updated unit workload version. This prevents creating a in memory
sort during reading the status for a unit.