feat: add support to DDL for agent stream in model
This commit changes the DDL for model agent to start tracking the agent
stream that is in use for a model. Previouusly this value would be
stored and tracked in the model config table.

It has been noticed that we need to represents this value as a first
class attribute in the model that can later be used by other components.

Because of the way this is introduced this single commit will fail a lot
of testing around insert statements. Follow up commits will introduce
fixes.