refactor: integrates agent stream into domains
With the addition of agent stream to the agent version table we now need
to plumb this value through the various places where it will be set.
This is so that the DDL changes made in the previous commit work and
don't cause test failures.

This commit makes no attempt to set the value correctly or use the value
it is just about making the code function and not having the new DDL
fail.

A follow up PR will do the work of wiring in the agent stream and making
sure that it gets read and set correctly.

As part of this commit I have also introduced a new core type to
represent the agent stream values previously found in
environs/tools/simplestreams.go