feat: remove pubsub metrics
We're removing pubsub, so we need to see what's actually left
dangling. The peergrouper and the controller config is know to
be left, but there is one more. The opening of the controller
port API is also a pubsub message. This is surprising as it's
possible to just watch the controller config for changes and open
the api port.