refactor: remove half implemented import code and add todo
Importing of Device constraints and application storage need to be
investigated. They are not fully implemented in the modelmigration
package. Remove the half done pieces to ensure it's clear what isn't
done yet in application migration.

Remove the commented out call to insertApplicationStatus as part of
application migration. This is done in the status domain.