feat: push status history into status domain
The status history reader is better served not being in the api
layer. That way we can marshal off the transport layer into the
the domain and back out. That way we can change the implementation
without changing the facade.

Test will be required still, but this gets us there.