feat: support setting the model's agent stream
This change introduces the concept to the model agent service layer for
setting the agent stream that is in use by the model. The state layer is
only defined at the moment and still needs to be implemented.

This is done so that the model config facade can intercept users setting
agent stream via model config and instead re-direct it to this service.