feat: adds support to model agent state for setting stream
This commit follows up with the state layer for adding support to set
and read the agent stream of the model. The service layer defined this
need in the last commit and this commit follows up with the
implementation.

The reading aspect of the service layer is currently used externally to
the package but it provides good testing utility. It is expected this
value will be used in follow up commits so it isn't wasted
implementation.