feat: set model agent stream from model config facade
This commit changes the model config facade so that the agent stream can
be extracted from the incoming model config and be used to set the agent
stream via the service. This allows model config to still act as an
interface for setting agent stream but we don't directly store the value
in model config anymore.