chore: make model config tests idiomatic Juju
This commit is a house keeping item so that testing for the model config
api calling the agent service can go again. What this commit does is
moving the testing code into the same package and deletes the mocks sub
directory favouring mocks in the same tree as the tests.

Also removed the secret suite as it makes no sense to be seperate in
this instance.