feat: model config agent stream validator
This commit adds a new validator to the model config domain for making
sure that the user cannot modify the agent stream value in model config
as it is now considered a read only value.

The validator will also remove the value from model config so that it c
an't be persisted into state.

As part of this commit I also removed the defaulting of the agent stream
value in model config. This is business logic and does not belong in
environ config.