refactor: set model agent stream in model config
When hydrating model config from the database we now pull the value of
the model's agent stream to also include in the model config. This
allows model config to keep being a readonly source for this value until
all of the places where it is read can be squashed out.