test: add tests for model config agent stream
This commit adds missing tests for agent stream in model config testing
that the value can't be changed or set in model config and persisted in
state.

As part of this change I also removed a test from the service for
watchers. We have a watcher integration test at the domain level and
don't require this anymore.

Also removes the default value for agent stream from environ config so
that we don't have business logic from environ config being inserted
outside of the domain layer.