test: add set model config test back in
In the previous commit I removed the set model config test as it was
more testing watchers then anything else. We don't require the watcher
tests as this is done via an integration test in the top level domain.

The set model config aspect of the test has been brough back in.