chore: organise model config facade params
This is a PR feedback task to refactor the members of the model config
facade so they are logically grouped together and also ordered. It
allows for the facade to be easily read and understood as it undergoes
change.

As part of this I also inverted the error out of new and instead placed
it into make for NewModelConfig.