refactor: model bootstrap local record creation
This commit introduces a new local record bootstrap func that allows
callers to either use the variant that supports setting a custom agent
stream or let the agent stream be set to a default.

This is useful for cases where a user bootstraps a controller and either
specifies a stream or doesn't.