refactor: agent bootstrap for agent stream key
This commit changes agent bootstrap to inspect the model config that has
been passed by the user at bootstrap time. If the model config has the
agent stream key set it is extracted and passed to the local record
creation op to be set along side the agent version.

If this has not been set by the user we use the default path that sets a
default value.