refactor: stream used when building local juju
During bootstrap we have this code that attempts to build local binaries
to upload into the bootstrapped machine. The building process attempts
to create a unique directory to place these tools using the agent stream
that is set in model config.

This worked previously because the default value for agent stream would
be released. We know longer set this value anymore to a default value.
As this information isn't used to affect stream information but just to
create a directory to place built binaries we default this to
"bootstrap" now.

This same pattern has been followed in the update controller usage of
the BuildAgentTarball code.