fix: ensure details are reported
When passing ?detailed=true to any of the health probes, make sure all
of the code paths that can return alternative errors also can give you
the details.
While digging, found some poor formatting, since by default it puts the
detail information on the same line as the error string, and the
NotImplemented prober was returning the error "probe not implemented not
implemented".