refactor: remove model uuid from mock state
In the mock state that is held for application offers we had a model
uuid that was being used. This is now removed in favour of one that is
on the base suite it self.

This is to help cleanup uses of Mongo and better identify what is still
left to do. Change was made as a direct result of PR feedback.