test: add default stream to upload tools
We have a testing fixture that uploads fake tools for integration
testing of commands. This has previously work because of the default
value model config applied to agent stream.

This change modifies the upload function to default the stream to
released if the string is empty. This way the same behaviour is
maintained. These tests will eventually get moved to bash tests.