refactor: readonly model info creation
This commit refactors one of the functions we have for making readonly
model information in the model database. The function was changed to now
accept an agent stream to use alongside the agent version for the model.

We also set  a default of released if no value is suppled by the normal
create model func.

A new error type was also introduced to represent the fact that the
agent stream supplied is not valid.