refactor: environs config test
This commit changes the testing we had for model config around default
values for agent stream. The test was asserting that if no value was
supplied for agent stream a default one of "released" would be set by
the model config code.

This has since been removed and an empty value here is fine. The test
now checks to see if it has been set by the user and then only checks to
make sure it is reflected in the resultant model config AgentStream
getter.