fix: tests that were using tools dir over stream
We had this broken concept in a lot of our tests around tools dir v that
of a stream when behind the covers the two were being conflated with
each in weird ways.

Whenever we talking about a tools dir we were really always talking
about a stream. Tools dir was a way to represents the middle path of the
path we were using when this value was always the stream as well.

So what this PR does is get rid of tools dir from all of the tests as
they were always setting it to the same value as that of the stream.