feat: remove pubsub from facades
This removes the pubsub hub from the facade interface along with
the publishing the controller config changes. The solution and
should have been the solution from the get go, was to watch for
controller config changes and update the appropriate call sites.

The only vestigial of pubsub in the apiserver is for the handlers
when dispatching messages between controllers. Once the peergrouper
is gone and we no longer dispatch topics, this can also go.