chore: fix TestSSHWorkerReport flaky test
The test was establishing a connection to the SSH server but not establishing an SSH session which made the test flaky.

The SSH server receives a connection, runs the connection callback, which increments the connection counter, then performs the SSH handshake. By creating a real SSH client we guarantee that the connection counter is incremented.