feat: wires up model status in common package
This wires up the status methods for model status commands in
package. There is a few barriers to jump over to get this to work
nicely. I'm not sold on the closures to get the services. This
puts a big strain on the domain services worker to constantly
find the services for the model. For me, this could be done just
once and then return the services for that one thing.