feat: handle context canceled when shutting down
Now that the apiserver has it's own watcher, this can be canceled
and the return statement from the catacomb wait can be context
canceled. This is considered a clean kill (for now), so we need to
just return nil when this happens.

The real fix is to set context cause to be either nil or some
sentienl error when the tomb is killed with nil.