feat: add ip address watcher by net node
This patch adds a new notify watcher, that notifies every time the ip
address changes for a given list of net node uuids.
To achieve this, the net_node_uuid column had to be added to the
ip_address table, and the trigger was updated to track changes to the
net_node_uuid column. This is needed to avoid hitting the DB on every
change to the ip_address table to filter by net node uuid, since this
table will change very often.
Also, a method for retrieving the net node uuid(s) of a given unit was
added, this will be needed to initialize the new watcher.
As a side change, a new predicate (`ContainsPredicate`) was added to
filter the events based on a list of values (needed in case we watch
more than one net node, which is the case in k8s).