feat(application): wire up three watchers in the uniter facade
Wire up the watcher: WatchConfigSettingsHash, WatchTrustConfigSettingsHash
and WatchUnitAddressesHash

In this wiring up, both the WatchConfigSettingsHash and
WatchTrustConfigSettingsHash are actually watching the same thing. This
is because trust is a field on the application_settings table, and the
watcher watchs only at the granulairty of some change within either the
settings or config table. This means the uniter will get over-notified
for changes to the trust table. But this should be ok. For this reason,
it may be the case that we can get rid of the trust config watcher
entirely, but I will leave this to be determined in the future.