fix: domain services bug in tests
To help with the transition of api server suite over to DQlite and to
fill the gap between full state testing and moving to mocks we developed
the domain service suite.

This suite is used in the same places the apiserver suite is in and
establishes a default model as part of the setup. The domain services
helper getter we had for the default model was incorrectly wired up to
the controller model.