refactor: credential state layer for invalidating
This is a small refactor of the state layer for credentials around
invalidating credentials. The main change is we now only accept a
credential uuid to invalidate and not the natural key.

This allows for the invalidation code to be used across common
implementations. The other change is around invalidating the cloud
credential that is in use by a model. This allows for a natural work
flow on the service when invalidations need to happen from the context
of a model.