feat: support for invalidating model credential
This commit updates the credential service to support invalidating the
credential that is in use by a model. This is required by the common
credential code and results in the facade not having to understand what
credential is in use by the model. The facade can now just use the model
information.

While here the service has been updated to use the new state layer where
a credential uuid is supplied downwards for invalidation.