chore: remove credential manager client facade
This commit removes the credential manager client facade. The juju
client callers were recently removed and the only place we have left
that calls this is 3.6 client and py lib juju.

For juju 3.6 it is the destroy controller command that calls this. We
don't support non 4.0 client destroying 4.0 controllers. I can't find
any current uses in pylib juju.

We will however be using a new pylib juju version when we release 4.0.